pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "personal-knowledge-platform"
version = "0.1.0"
description = "Reliability-first personal knowledge platform with fast and deep research runtimes."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "beautifulsoup4>=4.13.3",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "markdown-it-py>=3.0.0",
  "openai>=1.72.0",
  "pillow>=11.2.1",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.1",
  "pymupdf>=1.26.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.40",
  "trafilatura>=2.0.0",
  "typer>=0.15.2",
  "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
  "import-linter>=2.3",
  "mypy>=1.15.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.0",
]

[project.scripts]
pkp = "pkp.ui.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pkp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
strict = true
warn_unused_configs = true
packages = ["pkp"]

[tool.uv]
default-groups = ["dev"]
